fix bug

+++ lib/app/data/models/classification_result_model.dart
@@ -4,7 +4,7 @@ class ClassificationModel {
   File imageFile;
   String coralName;
   String accuracy;
-  int numberAccuracy;
+  double numberAccuracy;
   String areaId;
   String latitude;
   String longitude;
